add abd: 'arcanist branch daemon' WIP
To simplify interaction with Phabricator in some workflows, implement
a daemon that watches for branches matching a particular pattern,
e.g. "ph-review/description/base/to/diff/against"

When a matching branch is encountered, create a review in Phabricator
using the commit messages similar to how Arcanist does.  When the
branch is updated, also update the review.  When the revision is
accepted, automatically land the branch.

The first two parts of the branch name are fixed, i.e. it must begin
with 'ph-review' and follow with a single-word description, remaining
terms are assumed to be the branch to base the review on (and land on).

This first version has no tolerance for errors and is pretty much
implemented all in one file, will split it out into multiple modules.

Test Plan:
Implemented 'TestAbd' to exercise the no-errors case