phlsys_git, phlgit_*: abstract usage of git
Add a 'Clone' class to keep track of working directory for operations
on a particular clone.  Add modules for each git subcommand that we're
interested in.  Separating the 'Clone' from the methods means that it's
easier to mock a 'Clone' object for testing the subcommands and clients
thereof.

Test plan:
the modules have unit-tests associated with them

Reviewers: jevripio
Differential Revision: https://sibld.phab.dev.bloomberg.com/D91