phlsys_conduit, phlcon_*: abstract conduit usage
Add a 'Conduit' class to keep track of uri, user for operations on
a particular Phabricator instance.  Add modules for each conduit API
that we're interested in.  Separating the 'Conduit' from the methods
means that it's easier to mock a 'Conduit' object for testing the
APIs and clients thereof.

Test plan:
the modules have unit-tests associated with them

Reviewers: jevripio
Differential Revision: https://sibld.phab.dev.bloomberg.com/D93