initial.db: use @server.test for test user emails
We don't want to accidentally be sending emails to fake.com,
use server.test instead.

Note that the top-level-domain test is actually reserved by the
Internet Engineering Task Force in RFC 2606 as a domain name that
will never be used on the internet for production use so it is
safe for emails to accidentally be sent there.

Test Plan:
$ /phabricator/instances/dev/phabricator/bin/storage destroy --force
$ sudo puppet apply vagrant/puppet/phabricator/manifests/default.pp \
    --modulepath vagrant/puppet

Browsed to user email for Alice -> updated to new email address