abdcmd_default: exception type for bad commit msgs
Raise the new CommitMessageParseException if Phabricator doesn't
like the message digest that we send it.  The intention will be to
email the user with appropriate instructions to remedy the situation
and create a 'bad' working branch to monitor the situation.

Test plan:
Added 'test_badMsgWorkflow' to exercise this code path
./runtests