introduce common prefix for all bad branches
To make it easier to work with bad branches, split them into 3 types
sharing a common prefix.  The types signify the stage in the pipeline
and the next required action from the daemon when the branch is updated
by the user.

Fix untested bug in updateReview - only verify the branch when it has
been updated to prevent spam.

Test Plan: $ ./runtests