retire env magic - not needed with nose?
flake8 doesn't like the 'unused import' of env, I accidentally
removed it thinking it was an unused system import.  When reviewing
the change before committing it, I noticed what it was for again but
it didn't seem to be needed when running the tests with nose.

Apparently PEP8 'highly discourages' relative imports; I'm not really
sure what the best way of portably making phlsys_subprocess available
is, I guess it need to be on the PYTHONPATH but that seems a bit
awkward - ideally the thing could be run from the directory without
requiring extra installation steps.
(which is why everying else is dumped in the parent dir atm)