move from tests/ into 'main dir' / dumpng ground
Adopt approximately the BDE naming convention for test suites - use
a 't' suffix.  For BDE typically '.t.cpp' would be used, in Python
the equivalent might be '__t.py'.

Underscore is used because otherwise nose gets confused and tries to
import 't'.  It may be possible to avoid that with configuration but
maybe it's easier to change the convention from the start.

Double underscore is used to further distinguish the t from the rest
of the name.

Test Plan: $ ./runtests