arcyd: 'single' and 'multi' cmds, config files
To make Arcyd usable for large numbers of instances and repositories,
split the command-line into 'single' and 'multi' variants.

The 'single' variant retains the original interface and can be used for
easier testing and playing around.

The 'single' variant additionally supports loading of parameters from
one or more files by using the '@' prefix to specify filenames.

e.g.
$ arcyd single @my_param_file1 @my_param_file2 --sleep-secs 10

The 'multi' variant manages multiple instances of the 'single' variant
by allowing the user to specify files to load the configs from.

e.g.
$ arcyd multi --repo-configs @file1 @file2 --repo-configs @filea @fileb

Add an arycd_test/ directory with artifacts for testing out the single
and multi commands with the 'instant Phabricator' config that comes
with 'phabricator-tools'.

Test Plan:
$ ./runtests
Ran the single and multi commands in the ./arcyd_test directory