make empty revisions when failing to create revs
In order to move the discussion of a branch to Phabricator sooner,
make an empty review in cases where we can't create a branch in the
normal way successfully.

Add a generic exception commenting function to the Commentor so that
it can do the differentiation between exceptions itself as appropriate.

updateReview() will now retry bad branches even when there are no new
changes, in case a change in the environment means that it is now valid
(e.g. registered users, the base branch)

updateReview() can now update a review and land it in one go.

_phabUpdateWithExpectations runs processrepo 3 times to make sure that
the expectations are met each time, making processrepo idempotent.

test_mergeConflictWorkflow only needs to update the repo once at the
end since we can now update and land in one go.

Add abdt_conduitgit.getAnyUserFromBranch for a more robust
identificaton of someone who can be responsible for the revision.

Add abdt_gittypes.makeGitWorkingBranchFromParts() for convenience.

abdt_workingbranch.pushStatus() now returns the updated working_branch
so it can be used in subsequent operations.

Test Plan: $ ./runtests