use an embarrassing default test plan if omitted
It's probably going to be a common newbie mistake to leave the
test plan field out; given that arcyd isn't nearly as interactive
as something on the command-line, it's probably friendliest to use
an embarrassing default test plan which people will happily change
when they know how.

Also automatically comment on the revision with git-based
instuctions on how to remedy the situation which use the actual
branch name so they are hopefully copy-paste.

Test Plan:
updated BadMsgWorkflow test, $ ./runtests
inspected resulting review page.