use 'types' of sendmail binary for '-t' option
'sendmail' and 'catchmail' aren't entirely compatible and require
different parameters to get into the "all input from stdin" mode
that we want.

Allow configuration of --sendmail-type from the command-line which
in turn will use the correct parameters for the particular type.

We use the notion of 'type' instead of specifying the parameters
explicitly because it becomes complicated to allow for parameters
to another program when parsing our own parameters.

Test Plan:
in arcyd_test/ run maily test and single, multi Arcyd tests
$ ./runtests