arcyd: retry after delay in case of exception
It seems to be the case that if we run 'git fetch' in our environment
enough times then it will fail on authentication; subsequent attempts
will succeed.

As a stop-gap measure; allow Arcyd to retry at the top level a couple
of times before finally exiting.

Test Plan:
tried in single testbed with very small delays, added validation
routine for the delays.