phlsys_strtotime: make parsing times easier
To make time input a little easier; write a simple layer on top of
datetime.timedelta to convert e.g. '2 hours 10 minutes' to an actual
timedelta.

There's an awesome funcion in PHP for this sort of thing which accepts
natural language 'strtotime', this is where the component name comes
from.

Test Plan:
add doctests, $ nosetests py/phlsys_strtotime.py --with-doc