arcyon comment: add 'actAsUser' support
Sometimes it's necessary to fix individual revisions as admin,
for some things (like abandon) you might want to act on the behalf
of another user.

N.B. you could commandeer the revision and then perform the action
but for some things it might make more sense to have the original
author appear to do it.

Test Plan: play in testbed/, abandon a revision for another user