arcyd multi: --kill-file --reset-file --pause-file
To make the arcyd process a little easier to deal with, add
support for watching the filesystem for 'special files' that
will kill, reset or pause the process.

Reset here means resetting this list of active operations, as
some of them may have failed too many times and run out of
'retries'.

Test Plan: play in testbed/