add phlsys_namedtuple: helper for unstable schemas
Add a wrapper around collections.namedtuple which will allow us to
easily marshall dictionaries into namedtuples with the flexibility to
specify required, default and ignored parameters.

This is necessary to provide some defined flexibility in the schemas of
the data that we receive from Phabricator.

Test Plan: $ ./run_all_tests.sh