phlsys_namedtuple: warnings.warn on unexpeced arg
If an unexpected arg is passed to the namedtuple factory then emit a
warnings.warn and ignore the unexpected args.  This allows the
namedtuple to be robust in the presence of new unhandled attributes and
allows the owner of main() to decide how to handle the warnings.

This also fixes the broken behaviour in the presence of unexpected
params if asserts are disabled - it would continue only to have the
collections.namedtuple raise an exception.

Test Plan:
Updated unit tests to expect warnings.warn's
$ ./run_all_tests.sh