Arcyd: impose a limit on the size of patches
If a diff is larger than the imposed limit then try to reduce it by
reducing context.  If reducing the amount of context doesn't help then
try it without context.  If this still hasn't reduced the size of the
diff enough then generate an error and report it in the review.

Test Plan: added test_createHugeReview, test_hugeUpdateReview
$ ./run_all_tests.sh