gen package group doc from module docstrings
Add a new script to generate markdown files that catalog modules and the
first line of their docstrings as a way to inventory package groups.

Commit the generated documentation, generally we don't want to commit
anything that is generated.  In this case we don't trust the generator
and want to version the output.  Also we want the output to exist
alongside the source and be served up by GitHub.

May reconsider committing the documentation later, it generally seems
like more doc is a good thing though.

Create new precommit script for tests and docgen, create new
runtime_tests.sh for running nose

Test Plan: $ ./precommit.sh