Arcyon examples assume arcyon installed on PATH
Summary:
So that the Arcyon example scripts can be invoked from a directory other
than the one that they live in; change them to invoke arcyon directly.

This assumes that Arcyon is installed on PATH, as is documented by
`examples/README.md` and the `_check_requirements.sh` script.

Test Plan:
Installed arcyon on PATH
Created a new directory, copied a valid .arcconfig into it
Ran all of the changed scripts using relative paths, except 'nudge'

Reviewers: druoso

Reviewed By: druoso

CC: phab

Differential Revision: https://sibld.phab.dev.bloomberg.com/D127