add up-front version checks for 2.7 to executables
Summary:
The code base currently depends on version 2.7 of Python, any earlier
than that and it won't have the requisite argparse feaures.  Any later
than that (3.x) and there are breaking changes in the syntax.

Prevent nasty runtime surprises by enforcing version 2.7 as early as
possible.

    The version check itself will not work prior to Python version 2.0,
    that's when sys.version_info was introduced.

Test Plan:
Changed version requirement to earlier and later to ensure that it
fails both ways.

Reviewers: oparanal

Reviewed By: oparanal

CC: phab

Differential Revision: https://sibld.phab.dev.bloomberg.com/D139