Enhanced tracebacks, with local variables
Add a new component, phlsys_traceback which has support for printing
enhanced tracebacks with more information for debugging purposes.

Test Plan:
$ ./precommit.sh

Try installing the exception handler in various existing apps to see
that the output is as expected.  Some of the traces are a little
strange, possibly owing to outputting the contents of local variables.
Nothing appears to be actually wrong however.