Arcyd: handle branches with only empty commits
In the event that we encounter a branch comprised only of empty commits
then we will generate an empty diff, which Phabricator will reject.

Catch this early by checking for an empty diff and raising an
AbdUserException so that a failed review page will be generated.

Add a new 'test_emptyCommit' to exercise the creation case.

Test Plan: $ ./precommit.sh