Arcyd: extract conduitmock internals to new class
Make it possible for abdt_conduitmock.ConduitMock to have the same
interface as abdt_conduit but also allow us to meddle with its workings.

Split the internals of the mock out into a separate class, which
ConduitMock will delegate all of its reposonsibilities to; leaving it as
merely a wrapper for the internals.

We may then supply the internals class to the wrapper and do as we wish
with the internals, so long as we retain a reference to it.

Test Plan: $ ./precommit.sh