fix: handle error when updating closed revision
Raise a 'phlcon_differential.UpdateClosedRevisionError' in the event
that this should happen, which abdt_conduit will convert into an
abdt_exception as appropriate.

Update tests to assert that Differential won't allow updates to a
revision that has already been closed.

Update the conduitmock to replicate this behaviour.

Test Plan: $ ./precommit.sh