Test python file read/write semantics
Add testbed scripts to exercise the 'open' command and see if it behaves
the way we want in order to support sharing status files between Arcyd
and a websever.

Looks like it does what we want - readers only see the file in coherent
states, the writer doesn't appear to be starved and unable to write.

Test Plan: $ ./precommit.sh