Added a trusted-plugin system
Summary:
A system for specifying plugins which do not get copied versions of the references passed to the plugin in params.

If plugins are loaded with --trusted-plugins then the references are not copied and any exceptions generated by the plugin will not be caught by the plugin manager.

Plugins loaded with --plugins now have extra seperation from arcyd.

Test Plan: Working on it

Reviewers: jevripio, phab

Reviewed By: jevripio

CC: phab

Differential Revision: https://sibld.phab.dev.bloomberg.com/D169