Arcyd: mention repository and branch url on create
When creating a new review, mention the repositiory that it was created
from and the URL to browse the contents of the branch (if configured).

This makes it easier to determine the origin of a change and to get the
full context when a review is too big to be fully displayed in
Phabricator.

Store the new information in the abdt_branch, test that both branch and
branchmock store and retrieve correctly.

Add link formatting support to abdt_remarkup.

Test Plan: $ ./precommit.sh