Arcyd: remove abdi_processrepo old tests
abdi_processrepo__t.Test is now on a par with the 'old tests' and seems
to have the same coverage.  The new style tests use mocks exclusively
so it's easier to set expectations and adding tests doesn't increase

Remove the 'old tests' and the code that is consequently no longer used.

Test Plan:
Deliberately try to break processrepo__t and ensure that the new tests
detect the fault.
$ ./precommit.sh