Arycd: get landing author from latest commit
Unconditionally copy the author's details from the latest commit
on the review branch and use those for the landing commit.

This replaces the 'determine primary Phabricator user from branch' logic
which is less robust.

Add a new test for the 'latest committer is not a user' scenario.

Test Plan: $ ./precommit.sh