add .travis.yml to run static_tests.sh
Travis CI is a hosted continuous integration service for the open
source community.

We need to add a configuration file to tell it how to provision a VM
that can run our tests, and how to run our tests.

Test Plan: use http://lint.travis-ci.org/ to validate the file