Arcyd: add 'arcyd-status-html' command
Add a new command for rendering a dashboard page for the running Arcyd
instance.

Move the existing 'status-html' command to 'repo-status-html'.

Introduce a new 'arcydreporter' for capturing the relevant state of the
running instance.  In later work we can move everything from stdout and
stderr into this mechinism, it should probably capture standard logging
too.

Extend devstatushtml to cover the new command and exercise the
'arcyd-status-html' command in the system tests.

Test Plan: $ ./precommit.sh