Arcyd: add more detail to 'no users' error
Add more detail to the 'Unable to assign any users to branch.' error
message that is displayed on the dashboard, better to suggest some
things for people to copy-paste then just refer them to the docs.

Test Plan:
$ ./testbed/arcyd/test_shell.sh
manually create a review with bad author, it turns out that '<text>'
won't appear on the dashboard as it's getting converted to a tag, will
deal with that separately.

$ ./precommit.sh