barc gc: automatically prune landed branches
To help users with the potentially fiddly config to fetch
'refs/arcyd/landinglog' locally, add a prompt if it doesn't appear to be
already set up.

Introduce phlsys_choice to help with presenting such menus to the user.

Don't set up the local repo to fetch the upstream landinglog if it
doesn't exist yet.

There's no point in setting up the local repo to fetch the landinglog
ref if it doesn't exist, in fact this can present the user with some
very puzzling errors.

Test Plan:
poke, accept and gc in
$ ./testbed/arcyd/test_shell.sh
$ ./precommit.sh