Arcyd: fix exception when landing msg not ascii
Fix an exception that occurs if the commit message suggested by
the Phabricator server for a particular review by the
'get commit message' conduit API cannot be converted to ascii.

Introduce a new module to lossily convert to ascii from unicode and use
that to post-process the commit message from Phabricator.

Test Plan:
poked in test plan with 'horizontal ellipses', exercised with shell:
$ ./testbed/arcyd/test_shell.sh

$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh
$ ./testbed/arcyd/loop_test.sh