arcyd: fix abort when author is also reviewer
Phabricator does not normally permit the author of a review to be a
reviewer, enforce this rule in Arcyd by removing the author from the
list of reviewers before sending the list of reviewers to Phabricator.

Notify the user that this has happened so that they may check for other
errors in the list of reviewers.

Test Plan:

added coverage to
$ ./testbed/arcyd/acceptance_tests.sh
reviewed error message to make sure it's suitable.

$ ./precommit.sh