phlgit_log: add 'abbrev_hash' field to Revision
Widen phlgit_log.Revision to include an 'abbrev_hash' field.

This is useful wherever we're displaying hashes on screen as typically
we only need the hash which disambiguates the commit now as opposed to
'in all likelyhood for all time'.

Shorter hashes mean we can fit more on the screen and there's less
"noise" to filter out.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh