Arcyd: don't care if can't push landinglog
There are repos where we can't push the landinglog due to permissioning
and won't be able to for some time.  Don't stop the show in this case.

Later we'll want to nag on the review to change the permissions or at
least advise that the landinglog doesn't contain this review yet.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh