Arcyd: better comment on 'missing base' reviews
When a user pushes a branch with a non-existant base, explain how the
name of the branch was interpreted to derive the name of the base and
list out the next possible steps.

Provide a one-liner for deleting the errant branch.

Test Plan:
$ ./precommit.sh
read the 'bad_base' review output from:
$ ./testbed/arcyd/acceptance_tests.sh