arcyon: print summary by default on create/update
Summary:
when using 'arcyon create-revision' or 'arcyon update-revision'
then print a human-readable summary of the operation and the link
to the review so that general users don't have to know how to convert a
revision id into a url.

Provide a new command-line option '--format-id' to get the old
behaviour.

Test Plan:
$ ./precommit.sh
updated and ran
$ ./testbed/arcyon/exercise_arcyon.sh

Reviewers: oparanal, raiyenga, phab

Reviewed By: phab

CC: arcyd

Maniphest Tasks: T33

Differential Revision: https://sibld.phab.dev.bloomberg.com/D171