Arcyd: refactor, switch to ClassicNaming
Replace usage of abdt_naming.makeWorkingBranchFromName with
ClassicNaming.make_tracker_branch_from_name and remove the redundant
function.

Explicitly instantiate ClassicNaming where it's needed, later we'll pass
it to where it's needed and introduce other types of naming.

We'll repeat this process with the other free functions of abdt_naming.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh