arcyon showconfig: clearer output
Summary:
Change the language of 'changed uri' to 'assumed uri', otherwise it can
seem as if it has modified a config file.

Make the obscured display of certs consistent by extracting a new helper
function.  'Top and tail' the cert for easier recognition that the value
is as expected without revealling the whole thing.

Test Plan:
$ arcyon show-config --uri http://127.0.0.1
$ ./precommit.sh

Reviewers: jwilkin1, rthomps1, jevripio

Reviewed By: jevripio

CC: arcyd

Differential Revision: https://sibld.phab.dev.bloomberg.com/D185