gen_doc: generate manual pages under doc/man/*
Make it easier to access the documentation for each Arcyon command by
writing the output of 'arcyon COMMAND -h' to a versioned file.

This has the very convenient side-effects of
- assigning each command reference a URL on github.com
- making changes to command doc easier to review
- bulking out the documentation to make the whole thing look better

This has the not so convenient side-effects of
- taking up a bit of extra space in git history and working copy
- complicating merges slightly, need to be careful to gen-doc for the
  correct result
- contributors may be tempted to edit the .txt directly with
  disappointing results

On balance it seems like a good thing, could also consider using
a gh-pages branch for this but that wouldn't have all the same
advantages and be a bit fussier to update.