proto/linterate: an automated lint excitor
Download files from differential reviews and apply robust linters to
those files.  Where almost certain errors are detected, make inline
comments on those reviews.

This is emphatically not a stand-in for CI, more a utility for getting
developers excited about CI and showing what's possible with some of the
fantastic linters out there.

This improves on 'phlint' by adding configuration options and a memory
of what it's already covered.  'linterate' only supports cppcheck for
now, 'pyflakes' and 'jshint' to follow as they become battle-tested.

Test Plan:
$ ./precommit.sh
linterated a large instance.