Arcyd: quickfix: wrong authors assigned to commits
phlgit_log.get_author_names_emails_from_hashes() will return a list of
names and authors.  Each unique entry only appears in the list once, at
the first time it's mentioned in the input list.

Change the order that we input hashes to this function, such that it's
most recent commit first.

Otherwise, if the most recent author appears earlier in the list then
that author will not be considered.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh

Contrive a situation where the most recent author wouldn't be
considered, e.g.

  Alice - most recent
  Bob
  Alice

Verify that Alice would not be author before the fix and would be after,
using
$ ./testbed/arcyd/test_shell.sh