phlsys_tryloop: update make_default_endless_retry
Change the design between make_default_endless_retry() such that it's
possible to watch for 6 fails in any given 60 minutes and be fairly sure
that there's a serious outage starting or ongoing.

This makes it possible to set a simple alert rule for '6 times an hour'
in external monitoring tools to trigger or continue an alert.

Test Plan:
$ ./precommit.sh

Made sure the numbers add up.