Arcyd: add a simple way to safely fetch and prune
This safely works around the case where a branch that would be fetched
conflicts with a branch that has already been deleted, this case causes
a straight-forward 'git fetch -p' to fail.

e.g. given this sequence of events:

  alice $ git push origin HEAD:refs/heads/mybranch
  bob   $ git fetch -p
  alice $ git push origin :refs/heads/mybranch
  alice $ git push origin HEAD:refs/heads/mybranch/blah
  bob   $ git fetch -p

bob would get an error when fetching the last time and would have to
'git remote prune origin' before fetching.

Test Plan:
$ ./precommit.sh

run through the example above, with arcyd playing the part of bob
also ran through manually as a separate dev as bob
$ ./testbed/arcyd/test_shell.sh