Arcyd: revert me, combine archive pushes
Temporarily combine all of the three pushes required to archive an
abandoned branch to reduce pushing overhead.

The reduced pushing overhead will clear the abandoned branch backlog in
roughly a third of the time (~3 hours -> ~1 hour).

This means that they can't be pushed in the order that guarantees
either success or no change.

This change will be reverted when the bulk of the abandoned branches
have been archived.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh

set 'recently updated' review threshold low and abandon branches
$ ./testbed/arcyd/test_shell.sh