Arcyd: reduce extremely larges changes to diffstat
Some changes are too large to review with Phabricator, users will still
expect some sort of feedback about the nature of the change and may wish
to use Phabricator and Arcyd to 'approve' and land the change.

In the worst case, we don't want a daemon to get in the way of a change
being landed if a fully informed human has decided to do so.

Attempt to display a diff stat of the change instead of the real diff,
which is too large.

Test Plan:
$ ./precommit.sh

added new 'test_large_add' case to acceptance_tests.sh
$ ./testbed/arcyd/acceptance_tests.sh