Arcyd: add 'add-repo' command
Arcyd is hard to setup.

Make it a bit easier to get started by providing an 'add-repo'
command which creates an appropriate configuration file for the
specified parameters.

Clone the repository and test writing to some of the refs we expect to
manage with Arcyd.

Test Plan:

Use 'arcyd add-repo' where appropriate in
./testbed/arcyd/exercise_arcyd.sh

$ ./precommit.sh