phlsys_textconvert: add to_unicode() for safety
This will raise an exception: unicode(unicode('hello))

Sometimes we need to ensure something is unicode at more than one
place and 'unicode()' seems to be the way to do it.

Provide a 'to_unicode()' routine to safely ensure that the supplied
string is a 'unicode' builtin.

Test Plan:
added breathing test to test driver
$ ./precommit.sh