phlsys_textconvert: add ensure_ascii() for safety
When we get byte strings with non-ascii characters from external
sources, we run the risk of a conversion error when printing or
converting the string to unicode.

Provide an 'ensure_ascii()' routine to safely ensure that the supplied
string is a valid ascii byte string.

Test Plan:
extended breathing test
$ ./precommit.sh