Arcyd: fix repo paths for dashboard
Fix the parsing of the repo configs during 'arcyd start' to match the
expected input for 'arcyd process-repos'.

Fix the paths to the '.ok' and '.try' files to match the expected by the
reporter and dashboard code. In later work we'll want to make the
reporter and dashboard re-use the same code so the paths are identical.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh
$ ./testbed/arcyd/test_shell.sh

Conflicts:
	py/abd/abdcmd_start.py