Arcyd: BREAK, introduce 'arcyd add-repohost'
Add a new kind of config entity, the 'repohost', which allows
repositories to share configuration of:

- repo url
- repo snoop url
- branch url

This reduces error-prone repetition of tricky strings and makes it much
easier to change configurations later en-masse if necessary.

Make --repo-url a required argument of repositories, so it's possible
to introduce a '--repo-url-format' argument that the full repo url
can be derived from.

e.g.
  --repo-url myorg/myrepo
  --repo-url-format 'https://github.com/{}.git'
will expand to:
  "https://github.com/myorg/myrepo.git"

Pass the new --repo-url argument through the --branch-url-format
string as well so that can be re-used.

e.g.
  --branch-url-format
  'http://my.git/gitweb?p={repo_url}.git;a=log;h=refs/heads/{branch}'
would expand to
  'http://my.git/gitweb?p=myrepo.git;a=log;h=refs/heads/mybranch'

Remove --repo-snoop-url in favour of --repo-snoop-url-format, so that
can be re-used as above.

Convert tests to use new 'arcyd add-repohost' command.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/loop_test.sh
$ ./testbed/arcyd/test_shell.sh
$ ./testbed/arcyd/acceptance_tests.sh