Arcyd: BREAK, move '--arcyd-email' arg to init
Move the '--arcyd-email' argument to the 'arcyd init' command instead
of the 'arcyd add-repo' command.

This reduces redundant typing on the far more commonly called
'arcyd add-repo' command.

Add a new attribute to ArcydReporter so it can carry the email address
where it needs to go.  In later work it should actually use that address
itself.

Adjust existing tests to use the new interface.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/loop_test.sh
$ ./testbed/arcyd/test_shell.sh
$ ./testbed/arcyd/acceptance_tests.sh