Arcyd: BREAK, 'arcyd add-repo' has positional args
Make the add-repo command easier to use by changing phabricator,
repohost and url parameters to be positional, so that they don't have to
be prefixed with lengthy identifiers, e.g. '--phabricator-name'.

Add sensible defaults for 'repo_name' and 'repo_description' based on
the positional arguments.

This means that instead of an invocation like:

    $arcyd add-repo \
        --name localhost_fs_myrepo \
        --repo-desc 'localhost :: fs :: myrepo' \
        --phabricator-name localhost \
        --repohost-name fs \
        --repo-url origin

you can use a single line, like:

    $ arcyd add-repo localhost fs myrepo

which seems considerably easier.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/loop_test.sh
$ ./testbed/arcyd/test_shell.sh
$ ./testbed/arcyd/acceptance_tests.sh