Arcyd: move signal handler setup to arcyd start
To better ensure graceful exit, register the sigterm handler at the
start of Arcyd's start routing instead of during abdi_processrepos.

Note that arcyd instaweb doesn't seem to exit if it's processing when
sigterm is received - something is catching the exception.  If it wasn't
for that then we'd probably want to cover everything from Arcyd's main()
routine.

Test Plan:
$ ./precommit.sh