Arcyd: BREAK, daemonize 'arcyd start' by default
Users will generally want Arcyd to be a long-running process that isn't
tied to any particular terminal or session.  To that end, make
'arcyd start' launch as a daemon by default, with stdout and stderr
redirected to ./var/log in the arcyd dir.

Provide a '--foreground' option for preventing the daemonization, useful
during tests and if a third-party solution for daemonization is to be
used instead.

Test Plan:

Update existing tests to make use of new functionality.

$ ./precommit.sh
$ ./testbed/arcyd/test_shell.sh
$ ./testbed/arcyd/exercise_arcyd.sh
$ ./testbed/arcyd/acceptance_tests.sh