Arcyd: add 'fsck' command for fixing up instances
Introduce a command for checking the files in an Arcyd instance for
consistency and automatically fix some issues.

This is particularly useful for restoring the repositories under var/
to an Arcyd instance after cloning it from an Arcyd repository.
This is generally how you would restore an Arcyd instance from a
backup, since Arcyd stores all config in a local Git repo.

Propagate return values from Arcyd subcommands to the command-line so
'arcyd fsck' can exit with other than 0.

Test Plan:
$ ./precommit.sh

remove repos from /var and restore in
$ ./testbed/arcyd/test_shell.sh