Refactor, use __call__() instead of call()
There are a few classes in the code base which allow 'passing through'
of commands to another API or interface by supplying a 'call' method,
e.g.

    git_repo.call('status')

Python allows classes to be 'callable' by suppling a '__call__' method,
which is usable like so:

    git_repo('status')

This is shorter and seems to be just as clear as providing an explicit
'call' method.  It also has the benefit of making it a bit easier to
mock up such classes by making it possible to pass functions in their
stead.

Change phlsys_conduit, phlsys_repo and abdi_repo to provide '__call__'
instead of 'call'.  Fix up all existing references to '.call('.

All three classes are modified in one commit so it's easier to
mechanically replace usage and verify the changes, i.e.

  $ git grep '\.call('
  $ find py -iname '*.py' | xargs sed -i 's/\.call(/(/g'
  $ git grep '\.call('

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh
$ ./testbed/arcyon/acceptance_tests.sh