Arcyd cli: stop if working copy has changes
Prevent user error of using parts of the Arcyd cli which make commits
whilst there are staged or unstaged changes in the working copy.

This avoids the scenario where changes to auxillary files in the arcyd
repository are accidentally committed when adding new repositories.

Auxillary files might include e.g. helper scripts expected to with
enterprise-specific logic.

Test Plan:
check that this fails as expected
$ ./testbed/arcyd/test_shell.sh
    $  cd ../arcyd/
    $  arcyd stop
    $  rm config/repository/localhost_fs_origin2
    $  rm -rf var/repo/localhost_fs_origin2/
    $  touch blah; git add blah
    $  arcyd add-repo localhost fs origin2