Rename files in doc/man/arcyon, generate with loop
Simplify the routine to generate the arcyon doc text files to use a
loop.

Rename the arcyon doc text files so that they use a dash to separate the
words in multi-word subcommands. e.g. 'raw-diff' instead of 'raw_diff'.
This is more consistent and is easier to reduce to a loop.

Test Plan:
Ensure git reports that the files are renamed and have no content
changes.  'git rm' the doc files first to make sure they are all
re-generated.